+++ das-20200116/alibabacloud_das20200116/models/_describe_sql_log_records_response_body.py
@@ -198,7 +198,11 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
     def __init__(
         self,
         account_name: str = None,
+        affect_columns: str = None,
+        client_ip: str = None,
+        client_port: int = None,
         collection: str = None,
+        connection_id: str = None,
         consume: int = None,
         cpu_time: int = None,
         dbname: str = None,
@@ -212,10 +216,13 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         origin_time: int = None,
         parallel_degree: str = None,
         parallel_queue_time: str = None,
+        params: str = None,
         physic_async_read: int = None,
         physic_read: int = None,
         physic_sync_read: int = None,
+        protocol: str = None,
         return_rows: int = None,
+        row_key: str = None,
         rows: int = None,
         scan_rows: int = None,
         scnt: int = None,
@@ -234,8 +241,12 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
     ):
         # The account of the database.
         self.account_name = account_name
+        self.affect_columns = affect_columns
+        self.client_ip = client_ip
+        self.client_port = client_port
         # A reserved parameter.
         self.collection = collection
+        self.connection_id = connection_id
         # The amount of time that is consumed to execute the SQL statement. Unit: millisecond.
         self.consume = consume
         # The CPU execution duration. Unit: microsecond.
@@ -262,14 +273,17 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         self.parallel_degree = parallel_degree
         # The parallelism of the PolarDB for MySQL cluster.
         self.parallel_queue_time = parallel_queue_time
+        self.params = params
         # The number of physical asynchronous reads.
         self.physic_async_read = physic_async_read
         # The total number of physical reads.
         self.physic_read = physic_read
         # The number of physical synchronous reads.
         self.physic_sync_read = physic_sync_read
+        self.protocol = protocol
         # The number of rows returned by the SQL statement.
         self.return_rows = return_rows
+        self.row_key = row_key
         # The total number of rows that are updated or returned by the compute nodes of the PolarDB-X 2.0 instance.
         self.rows = rows
         # The number of scanned rows.
@@ -317,8 +331,20 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if self.account_name is not None:
             result['AccountName'] = self.account_name
 
+        if self.affect_columns is not None:
+            result['AffectColumns'] = self.affect_columns
+
+        if self.client_ip is not None:
+            result['ClientIp'] = self.client_ip
+
+        if self.client_port is not None:
+            result['ClientPort'] = self.client_port
+
         if self.collection is not None:
             result['Collection'] = self.collection
+
+        if self.connection_id is not None:
+            result['ConnectionId'] = self.connection_id
 
         if self.consume is not None:
             result['Consume'] = self.consume
@@ -359,6 +385,9 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if self.parallel_queue_time is not None:
             result['ParallelQueueTime'] = self.parallel_queue_time
 
+        if self.params is not None:
+            result['Params'] = self.params
+
         if self.physic_async_read is not None:
             result['PhysicAsyncRead'] = self.physic_async_read
 
@@ -368,8 +397,14 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if self.physic_sync_read is not None:
             result['PhysicSyncRead'] = self.physic_sync_read
 
+        if self.protocol is not None:
+            result['Protocol'] = self.protocol
+
         if self.return_rows is not None:
             result['ReturnRows'] = self.return_rows
+
+        if self.row_key is not None:
+            result['RowKey'] = self.row_key
 
         if self.rows is not None:
             result['Rows'] = self.rows
@@ -423,8 +458,20 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if m.get('AccountName') is not None:
             self.account_name = m.get('AccountName')
 
+        if m.get('AffectColumns') is not None:
+            self.affect_columns = m.get('AffectColumns')
+
+        if m.get('ClientIp') is not None:
+            self.client_ip = m.get('ClientIp')
+
+        if m.get('ClientPort') is not None:
+            self.client_port = m.get('ClientPort')
+
         if m.get('Collection') is not None:
             self.collection = m.get('Collection')
+
+        if m.get('ConnectionId') is not None:
+            self.connection_id = m.get('ConnectionId')
 
         if m.get('Consume') is not None:
             self.consume = m.get('Consume')
@@ -465,6 +512,9 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if m.get('ParallelQueueTime') is not None:
             self.parallel_queue_time = m.get('ParallelQueueTime')
 
+        if m.get('Params') is not None:
+            self.params = m.get('Params')
+
         if m.get('PhysicAsyncRead') is not None:
             self.physic_async_read = m.get('PhysicAsyncRead')
 
@@ -474,8 +524,14 @@ class DescribeSqlLogRecordsResponseBodyDataItemsSQLLogRecord(DaraModel):
         if m.get('PhysicSyncRead') is not None:
             self.physic_sync_read = m.get('PhysicSyncRead')
 
+        if m.get('Protocol') is not None:
+            self.protocol = m.get('Protocol')
+
         if m.get('ReturnRows') is not None:
             self.return_rows = m.get('ReturnRows')
+
+        if m.get('RowKey') is not None:
+            self.row_key = m.get('RowKey')
 
         if m.get('Rows') is not None:
             self.rows = m.get('Rows')